update chat completion

+++ crates/wormtongue/src/tongues/responses/openai/types.rs
@@ -5,11 +5,11 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Serialize, Deserialize, Clone, Default)]
 pub struct TokenCount {
     #[serde(default)]
-    pub input_tokens: u32,
+    pub input_tokens: i64,
 
     #[serde(default)]
-    pub cached_tokens: u32,
+    pub cached_tokens: i64,
 
     #[serde(default)]
-    pub output_tokens: u32,
+    pub output_tokens: i64,
 }
